requirements.txt
langchain_core==0.3.21
langchain_groq==0.2.1
langgraph==0.2.53
pytest==8.3.3
python-dotenv==1.0.1
streamlit==1.39.0
typing_extensions==4.12.2
